+++ src/main/groovy/com/trevorism/auth/model/Oauth2Tokens.groovy
@@ -2,10 +2,11 @@ package com.trevorism.auth.model
 
 class Oauth2Tokens {
 
-    String access_token
-    String id_token
-    String token_type
-    int expires_in
-    String refresh_token
+    String accessToken
+    String idToken
+    String tokenType
+    int expiresIn
+    String refreshToken
     String scope
+    String tenantId
 }