+++ src/main/groovy/com/trevorism/auth/model/SupportedOauth2Provider.groovy
@@ -1,5 +1,5 @@
 package com.trevorism.auth.model
 
 enum SupportedOauth2Provider {
-    Microsoft
+    Microsoft, Google
 }