+++ src/main/groovy/com/trevorism/auth/service/TokenService.groovy
@@ -14,5 +14,5 @@ interface TokenService {
     String issueToken(Identity identity, String audience)
     String issueRefreshToken(Identity identity)
     String issueInternalToken(Identity identity, String audience, String tenantId)
-    String issueTokenFromOauthProvider(SupportedOauth2Provider provider, Jws<Claims> claims)
+    String issueTokenFromOauthProvider(SupportedOauth2Provider provider, Jws<Claims> claims, String tenantId)
 }