+++ src/test/groovy/com/trevorism/auth/controller/GoogleControllerTest.groovy
@@ -1,5 +1,6 @@
 package com.trevorism.auth.controller
 
+import com.trevorism.auth.errors.AuthException
 import com.trevorism.auth.model.Oauth2Tokens
 import com.trevorism.auth.service.TokenService
 import com.trevorism.auth.service.oauth.Oauth2Parser
@@ -7,16 +8,18 @@ import io.jsonwebtoken.Claims
 import io.jsonwebtoken.Jws
 import org.junit.jupiter.api.Test
 
-class MicrosoftControllerTest {
+
+class GoogleControllerTest {
 
     @Test
     void testCreateToken(){
-        String microsoftIdToken = "ey.microsoft.token"
+        String googleIdToken = "ey.google.token"
         String fakeTrevorismToken = "ey.trevorism.token"
 
-        MicrosoftController controller = new MicrosoftController()
+        GoogleController controller = new GoogleController()
         controller.oauth2Parser = [parse: {tokens -> [:] as Jws<Claims> } ] as Oauth2Parser
         controller.tokenService = [issueTokenFromOauthProvider: {provider, claims, tenantId -> fakeTrevorismToken}] as TokenService
-        assert fakeTrevorismToken == controller.createToken(new Oauth2Tokens(idToken: microsoftIdToken))
+        assert fakeTrevorismToken == controller.createToken(new Oauth2Tokens(idToken: googleIdToken))
     }
+
 }